Q_strncpyz, Q_strcat: Revert back to using int as size argument
Using int is a valid design choice as long as you don't plan
on processing strings longer than INT_MAX and have means for error
handling when it's < 0.

id tech 3 has both (contrary to libc) and it used such interface from
the begining. Changing it without a full usage review caused some
subtle bugs.

+++ src/qcommon/q_shared.h
@@ -1087,8 +1087,8 @@ char	*Q_strupr( char *s1 );
 char	*Q_strrchr( const char* string, int c );
 
 // buffer size safe library replacements
-void	Q_strncpyz( char *dest, const char *src, size_t destsize );
-void	Q_strcat( char *dest, size_t size, const char *src );
+void	Q_strncpyz( char *dest, const char *src, int destsize );
+void	Q_strcat( char *dest, int size, const char *src );
 
 // strlen that discounts Quake color sequences
 int Q_PrintStrlen( const char *string, qboolean use102color );
